requirements.txt
pydantic==1.10.2
fastapi==0.83.0
fastapi-crudrouter==0.8.5
SQLAlchemy==1.4.41
uvicorn==0.18.3
requests==2.28.1
python-dotenv==0.21.0
passlib==1.7.4
python-jose==3.3.0
psycopg2==2.9.3
python-multipart==0.0.5
pytest==7.1.3
coverage==6.4.4